requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
dj-database-url==0.5.0
Django==1.11.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
fire==0.1.3
gast==0.2.0
grpcio==1.11.0
gunicorn==19.8.1
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.6
Markdown==2.6.11
numpy==1.14.3
Pillow==5.1.0
protobuf==3.5.2.post1
psycopg2==2.7.4
pytesseract==0.2.0
pytz==2018.4
PyYAML==3.12
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
whitenoise==3.3.1
